move get and has to superclass
